+++ examples/discovery-components-react-example/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import DiscoveryV1 from 'ibm-watson/discovery/v1';
+import DiscoveryV1 from '@disco-widgets/ibm-watson/discovery/v1';
 import './app.scss';
 import refinementsQueryResponse from './fixtures/refinementsQueryResponse';
 import { document as pdfDocument } from './fixtures/intro_to_watson_discovery';
@@ -26,7 +26,12 @@ const App = () => {
       searchClient={searchClient}
       collectionId={'3bbf3413-e6ee-a939-0000-016cc02a54c5'}
     >
-      <SearchInput />
+      <SearchInput
+        light={false}
+        small={false}
+        placeHolderText={'This is some placeholder text...'}
+        labelText={'This is some label text...'}
+      />
       <SearchResults />
       <SearchRefinements queryResponse={refinementsQueryResponse} />
       <ResultsPagination />