+++ packages/discovery-styles/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.1.1](http:///discovery-styles/compare/@disco-widgets/styles@1.1.0...@disco-widgets/styles@1.1.1) (2019-09-20)
+## [1.1.2](http:///discovery-styles/compare/@disco-widgets/styles@1.1.1...@disco-widgets/styles@1.1.2) (2019-09-23)
 
 **Note:** Version bump only for package @disco-widgets/styles
 
@@ -11,9 +11,12 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-# 1.1.0 (2019-09-19)
+## [1.1.1](http:///discovery-styles/compare/@disco-widgets/styles@1.1.0...@disco-widgets/styles@1.1.1) (2019-09-20)
 
+**Note:** Version bump only for package @disco-widgets/styles
+
+# 1.1.0 (2019-09-19)
 
 ### Features
 
-* add discovery-styles package ([#109](http:///discovery-styles/issues/109)) ([7b819ac](http:///discovery-styles/commits/7b819ac))
+- add discovery-styles package ([#109](http:///discovery-styles/issues/109)) ([7b819ac](http:///discovery-styles/commits/7b819ac))